Create Tkinter Framework
In this update using the tkinter framework the front end application requests for an image to select from the folder in it.
Once after the user gives the image, the image is read using opencv package. Then the image is downsized to 128*128 pixel.
Then the image is interplotted with the method inter_nearest to reduce the loss in image. Then this image is comapred to the model 
prediction from the model. If the value is 0 then it shows negative else positive.