training and testing of images
In the training function the values of start function are given as input. The parameters include positive training images, training negative
images, positive test images and negative test images. Once these images are arranged all the training and testing images are given to the model.
Here with input features of images respective class label is given as 1 or 0.