Create Convolutional Neural Network
With the template including preprocessing and training, model 1 is created using a Convolutional Neural Network
with multiple layers, where the input image is 128*128 pixels, the first layer include Conv2D layer with 32 nodes and
activation as relu. Then there is Maxpooling2D layer then with a Conv2D layer with 64 nodes fully interconnected, similarly
repeated twice to get max information and precision, and finally all the nodes are flatten to finally give 1 or 0.
Then this model is optimized with adam.