Create implementation of model
Here I implemented model where after all the preprocessing and training done, these values are given to the model,
here in this Convolutional Neural Network the input values is 128*128 pixels each time. In the first layer there is a
Conv2D layer with 16 nodes as hidden layers, then a MaxPooling2D layer with a consective Conv2D layer with 64 nodes
then with repeated 3 layers with 128,128 and 64 nodes respectively. Then finally the output is given to a flatten layer
with sigmoid function. The final output is either positive or negative.