Create KEERAS PACKAGE
In this model implementaion using Keras package the first layer includes 64 nodes with input size of 128*128 pixels.
Then the next layer is a maxpooling layer to reduce computations with maximum efficency. Then another convolutional
layer with 128 hidden nodes following with maxpooling, convolutional layers. In between a Dropout layer with 0.5 to 
inactive 50 percent nodes to reduce computation by minimum loss in data. And then finally a Dense flatten layer to 
get if the final value is either 1 or 0. Then these values are compared with the user value.